refactor: scope survey status action to list
